Adding autorization and removing unused routes

+++ app/Http/Controllers/CommentController.php
@@ -8,22 +8,11 @@ use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         //
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(CommentCreateRequest $request)
     {
         $comment = Comment::create([
@@ -36,37 +25,20 @@ class CommentController extends Controller
         return $comment;
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
-     */
     public function show(Comment $comment)
     {
         //
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, Comment $comment)
     {
         //
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Comment $comment)
     {
+        $this->authorize('delete', $comment);
+
         return $comment->delete();
     }
 
@@ -79,6 +51,8 @@ class CommentController extends Controller
 
     public function notSpam()
     {
+        $this->authorize('notSpam', Comment::class);
+
         $comment = Comment::findOrFail(request()->id);
         $comment->update(['spam_reports' => 0]);
         return ['spam_reports' => $comment->spam_reports];
